Moving binUtils to lib directory (src/bin)
